// generated/models/ApiMarketSendAttempt.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ApiMarketTransaction } from '../models/ApiMarketTransaction';
import { HttpFile } from '../http/http';

export class ApiMarketSendAttempt {
    'attempt_id': string;
    'purpose': ApiMarketSendAttemptPurposeEnum;
    'transaction_digest': string;
    'snapshot_block': number;
    'status': ApiMarketSendAttemptStatusEnum;
    'transaction': ApiMarketTransaction;
    'transaction_hash': string | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "attempt_id",
            "baseName": "attempt_id",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "purpose",
            "baseName": "purpose",
            "type": "ApiMarketSendAttemptPurposeEnum",
            "format": ""
        },
        {
            "name": "transaction_digest",
            "baseName": "transaction_digest",
            "type": "string",
            "format": ""
        },
        {
            "name": "snapshot_block",
            "baseName": "snapshot_block",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "ApiMarketSendAttemptStatusEnum",
            "format": ""
        },
        {
            "name": "transaction",
            "baseName": "transaction",
            "type": "ApiMarketTransaction",
            "format": ""
        },
        {
            "name": "transaction_hash",
            "baseName": "transaction_hash",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiMarketSendAttempt.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiMarketSendAttemptPurposeEnum {
    Approval = 'APPROVAL',
    Transaction = 'TRANSACTION'
}
export enum ApiMarketSendAttemptStatusEnum {
    Active = 'ACTIVE',
    Rejected = 'REJECTED',
    Resolved = 'RESOLVED'
}