// generated/models/ApiMarketReceiptTransaction.ts
export enum ApiMarketReceiptTransactionPurposeEnum {
    Approval = 'APPROVAL',
    Transaction = 'TRANSACTION'
}
export enum ApiMarketReceiptTransactionStatusEnum {
    Success = 'SUCCESS',
    Reverted = 'REVERTED'
}
export enum ApiMarketReceiptTransactionConfirmationEnum {
    Included = 'INCLUDED',
    Confirmed = 'CONFIRMED'
}